[Update] 농장 상세 조회, 수정 추가

+++ backend/src/main/java/org/example/groworders/domain/farms/repository/FarmRepository.java
@@ -5,12 +5,12 @@ import jakarta.validation.constraints.Positive;
 import org.example.groworders.domain.farms.model.entity.Farm;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface FarmRepository extends JpaRepository<Farm, Integer> {
+public interface FarmRepository extends JpaRepository<Farm, Long> {
     List<Farm> findByName(String name);
     Optional<Farm> findById(@NotNull(message = "농장은 필수 선택입니다.") @Positive(message = "농장을 확인 해주세요.") Long farmId);
+
 }
